refactor(be): podium 스키마 @toss/shared 이전 및 클라 응답 Zod 검증
