+++ client-toss/src/shared/api/serverToss.ts
@@ -1,4 +1,3 @@
-import type { PodiumResponse } from "@/entities/podium";
 import type { MyRankingResponse, RankingListItem } from "@/entities/ranking";
 import { appLogin } from "@apps-in-toss/web-framework";
 import type {
@@ -18,6 +17,7 @@ import {
   LoginResponseSchema,
   MyChanceResponseSchema,
   NotificationAgreementResponseSchema,
+  PodiumResponseSchema,
   PromptResponseSchema,
   SimilarityResponseSchema,
   SubmitDrawingResponseSchema,
@@ -188,8 +188,8 @@ export const serverTossApi = {
     return rankings;
   },
 
-  getPodium: (options?: RequestOptions) =>
-    get<PodiumResponse>("/rankings/podium", options),
+  getPodium: async (options?: RequestOptions) =>
+    PodiumResponseSchema.parse(await get<unknown>("/rankings/podium", options)),
 
   getDrawing: (drawingId: string, options?: RequestOptions) =>
     get<DrawingDetailResponse>(`/drawing/${drawingId}`, options),
