+++ packages/toss-shared/src/index.ts
@@ -4,4 +4,5 @@ export * from "./schemas/chance.schema";
 export * from "./schemas/archive.schema";
 export * from "./schemas/drawing.schema";
 export * from "./schemas/notification.schema";
+export * from "./schemas/ranking.schema";
 export * from "./schemas/user.schema";
